Main function used to produce a simulation

Main function used to produce a simulation

simulateTumour(
  birthrates = c(1, 1),
  deathrates = c(0.2, 0.2),
  mutation_rates = c(16, 16),
  clone_start_times = c(0, 256),
  fathers = c(0, 0),
  simulation_end_time = 1048576,
  seed = 42,
  number_clonal_mutations = 100,
  purity = 1,
  min_vaf = 0.01,
  depth = 100,
  depth_model = 2,
  verbose = FALSE,
  subset_fractions = numeric()
)

Arguments

birthrates
Birth rate(s) of subpopulations

deathrates
Death rate(s) of subpopulations

mutation_rates
Mutation rate(s) per division

clone_start_times
Timepoint (in number of reaction) a subpopulation is introduced

fathers
Subpopulation (0 based) a subpopulations originate from

simulation_end_time
Number of reactions performed

seed
Simulation seed

number_clonal_mutations
Number of clonal variants

purity
Assumed purity of the tumour

min_vaf
Minimum VAF to report

depth
Sequencing depth

depth_model
Number specificing the distribution of the sequencing depth to use (0: no sequencing, 1: poisson, 2: overdispersed beta binomial (default), 3: fixed)

verbose
Print progress?

subset_fractions
Optional, Numeric vector specificing fraction variants are subset to.

Value

A temulator_result_object.

Examples

simulateTumour()
simulateTumour(subset_fractions=c("WES"=0.03))